requirements.txt
# SkyMailr — Django email orchestration service
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg[binary]>=3.1
redis>=5.0
celery[redis]>=5.3
django-celery-beat>=2.6
dj-database-url>=2.2
httpx>=0.27
bleach>=6.1
pydantic>=2.6
Jinja2>=3.1
openai>=1.40
anthropic>=0.25
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
cryptography>=42.0

# Dev / test
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3
freezegun>=1.5
